LayoutTests/fast/events/touch/ios/block-without-overflow-scroll-scrolling-state.html

<!DOCTYPE html>
<html>
<head>
    <style>
        body {
            margin: 0;
        }
    </style>
    <meta name="viewport" content="initial-scale=1">
</head>
<body>
    <div id="eventTarget" style="overflow: scroll; height: 200px">
        <div style="height: 3000px"></div>
    </div>
    <p id="description">Verify that the block scrolling on overflow has touch event region with synchronous dispatch associated with it.</p>
    <pre id="output">
    </pre>
    <script>
        testRunner.dumpAsText();
        testRunner.waitUntilDone();

        function runTest()
        {
            let eventTarget = document.getElementById("eventTarget");
            eventTarget.scrollTop = 5;
            eventTarget.scrollTop = 0;
            document.getElementById("output").innerText = window.internals.scrollingStateTreeAsText();
            testRunner.notifyDone();
        }

        window.addEventListener('load', function() { setTimeout(runTest, 15); }, false);
    </script>
</body>
</html>
